Added testing for game class. Coverage is above 50 percent
